+++ .gitignore
@@ -11,6 +11,8 @@ node_modules
 dist
 dist-ssr
 *.local
+.env
+.next/
 
 # Editor directories and files
 .vscode/*
