+++ .nvmrc
@@ -1,1 +1,1 @@
-18.17.1
+25.2.0
