+++ src/components/PreviewCell.jsx
