+++ src/components/pages/pixel-galaxy.png
