+++ src/pages/_app.tsx
@@ -1,0 +1,12 @@
+import '../index.css';
+import type { AppProps } from 'next/app';
+import NavBar from '../components/NavBar';
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+    return (
+        <>
+            {/*<NavBar />*/}
+            <Component {...pageProps} />
+        </>
+    );
+}
