+++ src/pages/index.tsx
@@ -1,0 +1,5 @@
+import HomePage from '../components/pages/HomePage';
+
+export default function Index() {
+    return <HomePage />;
+}
