+++ src/utility/routes.js
@@ -3,7 +3,7 @@ export const HOME_PAGE_ROUTE = '/';
 export const PROJECT_PAGE_ROUTE = '/projects';
 
 // API
-export const API_URL = import.meta.env.VITE_BACKEND_API_URL;
+export const API_URL = process.env.VITE_BACKEND_API_URL;
 export const GET_PROJECTS_ROUTE = '/projects';
 export const GET_TOP_THREE_PROJECTS = '/top';
 export const GET_TAGS_ROUTE = '/tags';
